==> analysis/03b_Phylogenetic_Tree.Rmd <==
---
title: "Phylogenetic Tree Inspection and Rooting"
author: "Gurpreet Kaur"
date: "`r Sys.Date()`"
output: html_document
  toc: yes
  toc_float:
    collapsed: no
    smooth_scroll: yes
    toc_depth: 3
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, 
                      fig.align = "center",
                      fig.path = "../figures/03_Phylogenetic_Tree/") #send any figure output to this folder
```

#Before starting

##Set my seed
```{r set-seed}
# Any number can be chosen
set.seed(238428)
```


#Load packages
```{r load-packages}
pacman::p_load(tidyverse, phyloseq, ggtree, phytools, install = FALSE)
```


#Goals
1. Load the fasttree unrooted tree
2. Add tree to phyloseq object
3. Visualize and inspect tree with ggtree
4. Prune ASVs, if needed
5. Root our tree
6. Combine new tree with a phyloseq object
7. Save 2 phyloseq objects: 1. Unrooted tree phyloseq object, 2. Rooted tree phyloseq object


#Load data
```{r}
#preprocessed phyloseq object
load("data/02_PreProcessing/raw_preprocessed_physeq.RData")
raw_preprocessed_physeq


#load in the tree
unrooted_tree <- read.tree("data/03_Phylogentic_Tree/ASVs_unrooted.tree")
unrooted_tree

str(unrooted_tree)
```

#Merge phyloseq
```{r merge-physeq}
#Intiuition check
stopifnot(ntaxa(raw_preprocessed_physeq) == ntaxa(unrooted_tree))

#merge the tree with the phyloseq object
unrooted_physeq <- merge_phyloseq(raw_preprocessed_physeq, unrooted_tree)
unrooted_physeq
```

#Plot Tree with 'ggtree'
```{r plot-tree-unrooted}
#Make a basic tree
kingdom_node_tree <- ggtree(unrooted_physeq) +
  #color tips by kingdom
  geom_tippoint(mapping = aes(color = Kingdom)) +
  scale_color_manual(values = c("goldenrod1", "cornflowerblue", "grey")) +
  #add title
  labs(title= "Unrooted tree") +
  #add the node label
  geom_text(aes(label = node), hjust = 0.5, vjust = -0.3, size = 2) +
  #Move the legend to the bottom
  theme(legend.position = "bottom")
```

#Evaluate long branch
```{r eval-long-branch}
#view a specific clade
#Zoom in on origin tree: Node 2357

viewClade(kingdom_node_tree +
            labs(title = "Unrooted Tree: Node 2357"),
          node = 2357)

```
