==> analysis/04_Biodiversity.Rmd <==
---
title: "Biodiversity Analysis"
author: "Gurpreet Kaur"
date: "`r Sys.Date()`"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      fig.path = "../figures/04_Biodiversity")
```

#Set up the environment

##Set the seed
```{r}
set.seed(238428)
```


##Load libraries
```{r load-packages}
pacman::p_load(tidyverse, devtools, patchwork, iNEXT, phyloseq, install = FALSE)
```


##Load in the data
```{r load-data}
load("data/02_PreProcessing/raw_preprocessed_physeq.RData")
raw_preprocessed_physeq

#Intuition check
min(sample_sums(raw_preprocessed_physeq))

#setting colors for stations
station_colors <- c(
  "Shipping Channel" = "dodgerblue4",
  "Aransas Bay" = "dodgerblue2",
  "Copano West" = "#D9CC3C",
  "Copano East" = "#A0E0BA",
  "Mesquite Bay" = "#00ADA7")

# Make metadata dataframe
metadata_df <-
  raw_preprocessed_physeq %>%
  sample_data() %>%
  data.frame()

head(metadata_df)


```



#Goals

1. Calculate the Hill diversity of the samples
2. Evaluate the rarefaction curves
3. Evaluate the diversity values
4. Make notes of specific samples

#Diversity calculations with iNEXT
```{r calc-div}

#prepare input data
iNEXT_input_df <-
  raw_preprocessed_physeq %>%
  otu_table() %>%
  data.frame()

#check
dim(iNEXT_input_df)


#Run iNEXT to calculate hill numbers
#note that: species in rows and samples in columns
#remember to set the seed
iNEXT_data <- iNEXT(iNEXT_input_df, q = c(0,1,2), datatype = "abundance")


#save the file
save(iNEXT_data, file = "data/04_Biodiversity/iNEXT_data.RData")


```


#Evaluate the diversity
```{r load-div}
load("data/04_Biodiversity/iNEXT_data.RData")
iNEXT_data

str(iNEXT_data)
typeof(iNEXT_data)
```


# Plot Diversity 
```{r plot-rarefaction}
# Prepare Colors 
color_df <- 
  iNEXT_input_df %>%
  colnames() %>%
  data.frame()
# Check
head(color_df)
# Rename the column 
colnames(color_df)[1] <- "names"
# Check
head(color_df)


# Make a helper dataframe for plotting with colors 
iNEXT_color_df <- 
  color_df %>%
  # Fix the names for merging
  mutate(names = gsub(names, pattern = "[.]", replace = "-"),
         names = gsub(names, pattern = "X",  replace = "")) %>%
  # Merge with metadata
  left_join(metadata_df, by = "names") %>%
  # Merge with colors for plotting with ggiNEXT
  left_join(data.frame(station_colors = station_colors,
            station = names(station_colors)),
            by = "station")
```



# Make a helper dataframe for plotting with colors 
```{r}
iNEXT_color_df <- 
  color_df %>%
  # Fix the names for merging
  mutate(names = gsub(names, pattern = "[.]", replace = "-"),
         names = gsub(names, pattern = "X",  replace = "")) %>%
  # Merge with metadata
  left_join(metadata_df, by = "names") %>%
  # Merge with colors for plotting with ggiNEXT
  left_join(data.frame(station_colors = station_colors,
            station = names(station_colors)),
            by = "station")
```

# Plot Rarefaction with `ggiNEXT`

```{r ggiNEXT, fig.width=8, fig.height=3.5}
# Plot rarefaction! 
# rarefaction/extrapolation curve, type = 1 

# Order q: 
  # 0 = Richness/ Number of Total taxa
  # 1 = Exponential Shannon / Number of "Common" taxa
  # 2 = Inverse Simpson / Number of "Dominant" taxa 

ggiNEXT(iNEXT_data, type = 1, facet.var = "Order.q") + 
  facet_wrap(~Order.q, scales = "fixed") + 
  scale_color_manual(values = iNEXT_color_df$station_colors, guide = FALSE) + 
  scale_fill_manual(values = iNEXT_color_df$station_colors, guide = FALSE) + 
  scale_shape_manual(values = base::rep(17, nsamples(raw_preprocessed_physeq)),
                     guide = FALSE) + 
  theme(legend.position = "none")
```


#Manually plot the diversity
##Rarefaction
```{r iNEXT-manual}

iNEXT_manual_df <-
  iNEXT_data$iNextEst$size_based %>%
  dplyr::rename(names = Assemblage) %>%
  #Fix sample names
  mutate(names = gsub(names, pattern = "[.]", replace = '-'),
         names= gsub(names, pattern = "X", replace = "")) %>%
  #join with metadata
  left_join(., metadata_df, by = "names") %>%
  #Add colors
  left_join(., data.frame(station_colors = station_colors,
                          station = names(station_colors)),
            by = "station")
dim(iNEXT_manual_df)
str(iNEXT_manual_df)

#Plot it - rarefaction curve

iNEXT_manual_df %>%
  #filter out rows that are calculated by rarefaction from iNEXT
    dplyr::filter(Method == "Rarefaction") %>%
  #make the actual plot
  ggplot(aes(x = m  , y = qD, color = station, group = names)) + 
  #line
  geom_line() +
  #geom_point() +
           facet_grid(Order.q~station, scales = ) +
           scale_color_manual(values = station_colors) +
  theme(legend.position = "bottom")


```
