==> analysis/05_CommunityAnalysis.Rmd <==
---
title: "Between sample (Beta) Diversity of Microbes along a Salinity Gradient"
author: "Gurpreet Kaur"
date: "`r Sys.Date()`"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      fig.path = "../figures/05_CommunityAnalysis")
```

#Goals
1. Load in the phyloseq data with rooted tree
2. Evaluate sequencing depth and remove the sample
3. Normalize the read count between samples
4. Calculate community dissimilarities.
  a. Sorenson: all species?
  b. Bray Curtis: 
  c. Weighted UNIFRAC
5. Visualize the community data with two unconstrained ordinations
  a. PCoA
  b. NMDS


#Setup

##Load libraries
```{r load-libraries}
install.packages("vegan")
pacman::p_load(tidyverse, devtools, phyloseq, patchwork, vegan, install = FALSE)

#Load station colors
station_colors <- c(
  "Shipping Channel" = "dodgerblue4",
  "Aransas Bay" = "dodgerblue2",
  "Copano West" = "#D9CC3C",
  "Copano East" = "#A0E0BA",
  "Mesquite Bay" = "#00ADA7")
```


##Load data
```{r load-physeq}
#Load in rooted phylogenetic tree
load("data/03_Phylogentic_Tree/phytree_preprocessed_physeq.RData")
midroot_physeq_rm456
```


#Explore Read Counts

##Normalize read counts

#Calculate and Visualize community dissimilarity

##Sorenson PCoA

##Bray Curtis PCoA

##Weighted Unifrac PCoA

##Bray Curtis NMDS


#Test for statistical Significance with PERMANOVA and betaispR

#Session information
```{r session-info}
# Ensure reproducibility 
devtools::session_info()
```
